server/requirements.txt
Django==4.1.7
python-dotenv==1.0.0
spotube==0.5.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
